Modularize cache fill sharing and guard vary-sensitive reads
