+++ crates/rginx-http/src/cache/tests/lookup/background.rs
@@ -252,6 +252,7 @@ async fn cache_manager_lock_timeout_falls_back_to_bypass() {
             notify: Arc::new(Notify::new()),
             acquired_at_unix_ms: unix_time_ms(SystemTime::now()),
             generation: 1,
+            share_fingerprint: String::new(),
             reader_state: None,
         },
     );
@@ -285,6 +286,7 @@ async fn cache_manager_lock_age_allows_second_fill() {
             notify: Arc::new(Notify::new()),
             acquired_at_unix_ms: unix_time_ms(SystemTime::now()).saturating_sub(5_000),
             generation: 1,
+            share_fingerprint: String::new(),
             reader_state: None,
         },
     );
